# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: 命理工具 API
  version: 0.1.0
  description: Identity, profile, reading and presentation contract.
servers:
  - url: /
    description: Same-origin web entry
...
paths:
...
  /api/v1/account/referrals:
    get:
      operationId: listAccountReferrals
      tags: [Identity]
      summary: Read private referral campaign progress for the current account
      security:
        - deviceSession: []
      responses:
        "200":
          description: Safe referral progress without user or payment identifiers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountReferralsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/account/notifications/{notification_id}/read:
    post:
      operationId: markAccountNotificationRead
      tags: [Identity]
      summary: Mark one owned in-app notification as read
      security:
        - deviceSession: []
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - name: notification_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated safe in-app notification projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountNotificationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/readings:
    get:
      operationId: listReadings
      tags: [Readings]
      summary: List the newest Reading Versions owned by the current session
      responses:
        "200":
          description: Public Reading Version summaries only; payloads stay encrypted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/readings/{reading_version_id}:
    get:
      operationId: getReadingVersion
      tags: [Readings]
      summary: Poll the owner-scoped Reading Version state
      parameters:
        - name: reading_version_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Current Reading Version state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingStartResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    deviceSession:
      type: apiKey
      in: cookie
      name: mingli_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
        minLength: 32
...
  schemas:
...
    AccountReferralRewardResponse:
      type: object
      additionalProperties: false
      required: [status, occurred_at]
      properties:
        status:
          type: string
        occurred_at:
          type: string
          format: date-time
...
    AccountReferralCampaignResponse:
      type: object
      additionalProperties: false
      required:
        - campaign_key
        - version
        - state
        - starts_at
        - ends_at
        - per_inviter_limit
        - codes
        - invited_count
        - my_attribution_stage
        - rewards
      properties:
        campaign_key:
          type: string
        version:
          type: string
        state:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: [string, "null"]
          format: date-time
        per_inviter_limit:
          type: integer
          minimum: 1
        codes:
          type: array
          items:
            type: string
        invited_count:
          type: integer
          minimum: 0
        my_attribution_stage:
          type: [string, "null"]
        rewards:
          type: array
          items:
            $ref: "#/components/schemas/AccountReferralRewardResponse"
    AccountReferralsResponse:
      type: object
      additionalProperties: false
      required: [campaigns]
      properties:
        campaigns:
          type: array
          items:
            $ref: "#/components/schemas/AccountReferralCampaignResponse"
...
    ReadingVersionSummary:
      type: object
      additionalProperties: false
      required:
        - reading_version_id
        - reading_root_id
        - profile_version_id
        - capability_id
        - version
        - status
        - object_id
        - dimension_ids
        - horizon
        - created_at
      properties:
        reading_version_id:
          type: string
          format: uuid
        reading_root_id:
          type: string
          format: uuid
        profile_version_id:
          type: [string, "null"]
          format: uuid
        capability_id:
          type: string
        product_id:
          type: [string, "null"]
        runtime_capability_ids:
          type: array
          items:
            type: string
        version:
          type: integer
          minimum: 1
        status:
          type: string
          enum:
            - input_ready
            - waiting_input
            - terminal_stopped
            - prepared
            - completing
            - accepted
            - delayed
            - runtime_unknown
        object_id:
          type: string
        dimension_ids:
          type: array
          items:
            type: string
        horizon:
          type: object
          additionalProperties: false
          required: [kind_id, start, end]
          properties:
            kind_id:
              type: string
            start:
              type: [string, "null"]
              format: date
            end:
              type: [string, "null"]
              format: date
        prior_answer:
          type: [string, "null"]
        input_request:
          type: [object, "null"]
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    ReadingStartResponse:
      type: object
      additionalProperties: false
      allOf:
        - $ref: "#/components/schemas/ReadingVersionSummary"
...
    ReadingListResponse:
      type: object
      additionalProperties: false
      required: [readings]
      properties:
        readings:
          type: array
          items:
            $ref: "#/components/schemas/ReadingVersionSummary"
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, request_id]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
  responses:
...
    Unauthorized:
      description: The operation's required Device or Guest owner session is missing or invalid; an invalid submitted Device Session expires stale device and CSRF cookies
      headers:
        Set-Cookie:
          description: May expire stale mingli_session and mingli_csrf cookies before re-authentication
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: CSRF or origin validation failed
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: The resource does not exist or belongs to another owner
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...